쿼리 분리중
