+++ src/pages/Match/Search.tsx
@@ -1,62 +1,23 @@
-import React, { FC, useEffect, useCallback, useMemo } from 'react';
+import React, { FC, useEffect } from 'react';
 import { useInView } from 'react-intersection-observer';
-import { useInfiniteQuery } from '@tanstack/react-query';
 import { useNavigate } from 'react-router-dom';
-import { apis } from '../../apis';
-import { useDispatch, useSelector } from 'react-redux';
-import { RootState } from '../../redux/store';
-import { toggleSortShow, toggleSelectShow, toggleClear } from '../../redux/features/toggleSlice';
-import { subjectSearchShowClear, sortSearchShowClear } from '../../redux/features/searchSlice';
 import { RiArrowDownSLine } from 'react-icons/ri';
-import { categories } from '../../shared/constant/subjectTable';
-import { sortCategories } from '../../shared/constant/sortTables';
 import { Helmet } from 'react-helmet';
+import { useSearch } from '../../hooks/match/useSearch';
+import { useSearchList } from '../../hooks/queries/useSearchList';
 import LoadingSpinner from '../../components/Common/loadingSpinner';
-import { queryKeys } from '../../shared/constant/queryKeys';
 
 const SearchMatch: FC = () => {
   const navigate = useNavigate();
-  const dispatch = useDispatch();
   const { ref, inView } = useInView();
-  const sort = useSelector((state: RootState) => state.search.sort);
-  const subject = useSelector((state: RootState) => state.search.subject);
-  const categoryName = useMemo(() => categories.find((c) => c.value === subject), [subject]);
-  const sortName = useMemo(() => sortCategories.find((c) => c.value === sort), [sort]);
-
-  const fetchPostList = async (pageParam: number) => {
-    const res = await apis.getMainPostList(pageParam, subject, sort);
-    const data = res.data.content;
-    const last = res.data.last;
-    return { data, last, nextPage: pageParam + 1 };
-  };
-
-  const {
-    data: postList,
-    fetchNextPage,
-    isFetchingNextPage,
-  } = useInfiniteQuery([queryKeys.SEARCH, sort, subject], ({ pageParam = 0 }) => fetchPostList(pageParam), {
-    getNextPageParam: (lastPage) => (!lastPage.last ? lastPage.nextPage : undefined),
-  });
+  const { categoryName, sortName, handleToggleSelect, handleToggleSort, ClearAll } = useSearch('');
+  const { fetchNextPage, isFetchingNextPage, postList } = useSearchList('');
 
   useEffect(() => {
     if (inView) fetchNextPage();
   }, [inView]);
 
-  useEffect(() => {
-    return () => {
-      dispatch(toggleClear());
-      dispatch(sortSearchShowClear());
-      dispatch(subjectSearchShowClear());
-    };
-  }, []);
-
-  const handleToggleSelect = useCallback(() => {
-    dispatch(toggleSelectShow());
-  }, []);
-
-  const handleToggleSort = useCallback(() => {
-    dispatch(toggleSortShow());
-  }, []);
+  ClearAll();
 
   return (
     <>
