+++ src/pages/Match/keyword.tsx
@@ -1,34 +1,14 @@
 import React, { FC, useEffect } from 'react';
 import { useInView } from 'react-intersection-observer';
-import { useInfiniteQuery } from '@tanstack/react-query';
 import { useNavigate } from 'react-router-dom';
-import { apis } from '../../apis';
-import { RootState } from '../../redux/store';
-import { useSelector } from 'react-redux';
-import LoadingSpinner from '../../components/Common/loadingSpinner';
 import { Helmet } from 'react-helmet';
-import { queryKeys } from '../../shared/constant/queryKeys';
+import { useKeword } from '../../hooks/queries/useKewordList';
+import LoadingSpinner from '../../components/Common/loadingSpinner';
 
 const Keyword: FC = () => {
   const navigate = useNavigate();
   const { ref, inView } = useInView();
-  const keyword_ = useSelector((state: RootState) => state.keyword);
-  const keyword = keyword_['keyword'];
-
-  const fetchsearchList = async (pageParam: number) => {
-    const res = await apis.getSearchList(pageParam, keyword);
-    const data = res.data.content;
-    const last = res.data.last;
-    return { data, last, nextPage: pageParam + 1 };
-  };
-
-  const {
-    data: searchList,
-    fetchNextPage,
-    isFetchingNextPage,
-  } = useInfiniteQuery([queryKeys.KEYWORD, keyword], ({ pageParam = 0 }) => fetchsearchList(pageParam), {
-    getNextPageParam: (lastPage) => (!lastPage.last ? lastPage.nextPage : undefined),
-  });
+  const { searchList, fetchNextPage, isFetchingNextPage } = useKeword('');
 
   useEffect(() => {
     if (inView) fetchNextPage();
